Made a 'public' and 'private' API for plugins.
All methods stay overridable, but 'recomended override' methods are in the top layer, FinalOutpostPlugin.
