+++ Dictionary_Eclipse.txt
@@ -1,3 +1,4 @@
 configs
 plugin
 plugins
+config
